JS library updates to patch a security vuln
Github says serialize-javascript<2.1.1 has an XSS vulnerability. It's a
sub-dependency of a couple of the webpack-related libs, so I've updated
those. I wonder whether this is all still intercompatible! I wonder how
I would find out! :D D:

Had to replace uglify-webpack-plugin with terser-webpack-plugin; the
former is deprecated in favor of the latter.